Version08/requirements.txt
akasha-terminal
